EDKIIUS-26: MdeModulePkg/UsbBusDxe: Usb Fix
Fixed a bug which led to an ASSERT due to the USB device context being
maintained after a port reset, but the underlying XHCI context was
uninitialized. Specifically, Xhc->UsbDevContext is freed after a reset
and only re-allocates the default [0] enpoint transfer ring. In order to avoid a memory leak, device enumeration is performed after freeing the necessary buffers. This allocates the Xhc->UsbDevContext for all endpoints of the USB device. BZ 4456

Cc: Jian J Wang <jian.j.wang@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Cc: Hao A Wu <hao.a.wu@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Abner Chang <Abner.Chang@amd.com>
Change-Id: I41e08120f3a2b1b825229b1e93a2153bcc875e34
Signed-off-by: Britton Chesley <Brit.Chesley@amd.com>